+++ facile_rs/prepare_zenodo.py
@@ -9,6 +9,9 @@ This script creates an empty archive in Zenodo in order to reserve a DOI and a Z
 Both are stored in the CodeMeta metadata file provided as input and can be later used by the script ``create_zenodo.py``
 to populate the Zenodo archive.
 
+Optionally, the script can create a new version from an existing Zenodo record.
+See the option --zenodo-version-update for more details.
+
 Usage
 -----
 
@@ -18,43 +21,43 @@ Usage
     :prog: prepare_zenodo.py
 
 """
-
-import argparse
+import json
 from pathlib import Path
 
-from .utils import cli, settings
+from .utils import cli
 from .utils.metadata import CodemetaMetadata, ZenodoMetadata
 from .utils.zenodo import create_zenodo_dataset, prepare_zenodo_dataset
 
 
 def create_parser(add_help=True):
-    parser = argparse.ArgumentParser(add_help=add_help)
-    parser.add_argument('--codemeta-location', dest='codemeta_location',
+    parser = cli.Parser(add_help=add_help)
+
+    parser.add_argument('--codemeta-location', dest='CODEMETA_LOCATION',
                         help='Location of the main codemeta.json JSON file')
-    parser.add_argument('--zenodo-url', dest='zenodo_url',
+    parser.add_argument('--zenodo-url', dest='ZENODO_URL', required=True,
                         help='URL of the Zenodo service. Test environment available at https://sandbox.zenodo.org')
-    parser.add_argument('--zenodo-token', dest='zenodo_token',
+    parser.add_argument('--zenodo-token', dest='ZENODO_TOKEN', required=True,
                         help='Zenodo personal token.')
-    parser.add_argument('--dry', action='store_true',
+    parser.add_argument('--zenodo-version-update', dest='ZENODO_VERSION_UPDATE', default=None,
+                        help='Enable Zenodo version update. Can be "codemeta" or a Zenodo identifier. '
+                        'If omitted, a new Zenodo dataset is created without versioning. '
+                        'If set to "codemeta", a Zenodo identifier is searched in the CodeMeta file, and a new version is '
+                        'created from it (if found). '
+                        'Any other value is considered as a Zenodo identifier: a new version will be created from it. ')
+    parser.add_argument('--dry', action='store_true', dest='DRY',
                         help='Perform a dry run, do not upload anything.')
-    parser.add_argument('--log-level', dest='log_level',
+    parser.add_argument('--log-level', dest='LOG_LEVEL', default='WARN',
                         help='Log level (ERROR, WARN, INFO, or DEBUG)')
-    parser.add_argument('--log-file', dest='log_file',
+    parser.add_argument('--log-file', dest='LOG_FILE',
                         help='Path to the log file')
     return parser
 
 
-def main():
-    parser = create_parser()
+def main(args):
 
-    settings.setup(parser, validate=[
-        'ZENODO_URL',
-        'ZENODO_TOKEN'
-    ])
-
-    if settings.CODEMETA_LOCATION:
+    if args.CODEMETA_LOCATION:
         codemeta = CodemetaMetadata()
-        codemeta.fetch(settings.CODEMETA_LOCATION)
+        codemeta.fetch(args.CODEMETA_LOCATION)
         name = '{name} ({version}, in preparation)'.format(**codemeta.data)
     else:
         name = 'in preparation'
@@ -62,16 +65,30 @@ def main():
     zenodo_metadata = ZenodoMetadata({'name': name})
     zenodo_dict = zenodo_metadata.as_dict()
 
-    if not settings.DRY:
+    # Management of Zenodo versioning
+    old_zenodo_id = None
+    if args.CODEMETA_LOCATION and args.ZENODO_VERSION_UPDATE == 'codemeta':
+        # Check if CodeMeta file contains a Zenodo identifier
+        identifiers = codemeta.data.get('identifier', [])
+        if not isinstance(identifiers, list):
+            identifiers = [identifiers]
+        for identifier in identifiers:
+            if isinstance(identifier, dict) and identifier.get('propertyID') == 'Zenodo':
+                old_zenodo_id = identifier.get('value', None)
+    elif args.ZENODO_VERSION_UPDATE is not None:
+        # Use the provided Zenodo identifier to create a new version
+        old_zenodo_id = args.ZENODO_VERSION_UPDATE
+
+    if not args.DRY:
         # create Zenodo dataset
-        dataset_id = create_zenodo_dataset(settings.ZENODO_URL, settings.ZENODO_TOKEN, zenodo_dict)
-        dataset = prepare_zenodo_dataset(settings.ZENODO_URL, dataset_id, settings.ZENODO_TOKEN)
+        dataset_id = create_zenodo_dataset(args.ZENODO_URL, args.ZENODO_TOKEN, zenodo_dict, previous_version=old_zenodo_id)
+        dataset = prepare_zenodo_dataset(args.ZENODO_URL, dataset_id, args.ZENODO_TOKEN)
 
         doi = dataset.get('metadata', {}).get('doi', {})
         doi_url = 'https://doi.org/' + doi
 
         # Update Codemeta file with DOI and Zenodo ID
-        if settings.CODEMETA_LOCATION:
+        if args.CODEMETA_LOCATION:
             codemeta.data['@id'] = doi_url
             doi_entry = {
                 '@type': 'PropertyValue',
@@ -100,13 +117,15 @@ def main():
             else:
                 codemeta.data['identifier'] = [doi_entry, zenodo_entry]
 
-            Path(settings.CODEMETA_LOCATION).expanduser().write_text(codemeta.to_json())
+            Path(args.CODEMETA_LOCATION).expanduser().write_text(codemeta.to_json())
         else:
             print(dataset)
+    else:
+        print(json.dumps(zenodo_dict))
 
 
 def main_deprecated():
-    cli.cli_call_deprecated(main)
+    cli.main_deprecated(__name__)
 
 
 if __name__ == "__main__":